more runtime checks

+++ config/stage/bootup_stage.sh
@@ -2,9 +2,11 @@
 #!/usr/bin/sh
 
 # improve boot time by not waiting for ntp
-systemctl stop systemd-time-wait-sync.service
-systemctl disable systemd-time-wait-sync.service
-systemctl mask time-sync.target
+if [ -f /usr/lib/systemd/system/systemd-time-wait-sync.service ]; then
+  systemctl stop systemd-time-wait-sync.service
+  systemctl disable systemd-time-wait-sync.service
+  systemctl mask time-sync.target
+fi
 
 # enable ssh provision login -> disable root, allow provisioning account, password auth, use pam
 sed -i 's/^#\?PermitRootLogin.*/PermitRootLogin yes/' /etc/ssh/sshd_config
